ci: fix failing tests
