+++ mdformat_mkdocs/_helpers.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import re
 from functools import wraps
 from typing import Callable
 
@@ -23,3 +24,11 @@ def rstrip_result(func: Callable[..., str]) -> Callable[..., str]:
         return func(*args, **kwargs).rstrip()
 
     return wrapper
+
+
+def separate_indent(line: str) -> tuple[str, str]:
+    """Separate leading indent from content. Also used by the test suite."""
+    re_indent = re.compile(r"(?P<indent>\s*)(?P<content>[^\s]?.*)")
+    match = re_indent.match(line)
+    assert match is not None  # for pyright # noqa: S101
+    return (match["indent"], match["content"])
